+++ CHANGELOG.md
@@ -6,4 +6,5 @@
 - feature: new supported version for Minecraft 1.20.3!
 - feature: new supported version for Minecraft 1.20.2! (Alpha-Version)
 - fix: replace the MC_VER equality with in-equality.
+- fix: failed to find the class RunInitializer.